assignment 18 fini, touches Enter et Tab

+++ src/editor/view.rs
@@ -40,8 +40,9 @@ impl View {
             EditorCommand::Move(direction) => self.move_text_location(&direction),
             EditorCommand::Quit => {},
             EditorCommand::Insert(chararcter) => self.insert_char(chararcter),
-            EditorCommand::Backspace => self.backspace(),
+            EditorCommand::Backspace => self.delete_backward(),
             EditorCommand::Delete => self.delete(),
+            EditorCommand::Enter => self.insert_newline(),
         }
     }
 
@@ -53,9 +54,17 @@ impl View {
     }
 
     // region: Text editing
-    fn backspace(&mut self) {
-        self.move_left();
-        self.delete();
+    fn insert_newline(&mut self) {
+        self.buffer.insert_newline(self.text_location);
+        self.move_text_location(&Direction::Right);
+        self.needs_redraw = true;
+    }
+
+    fn delete_backward(&mut self) {
+        if self.text_location.line_index != 0 || self.text_location.grapheme_index != 0 {
+            self.move_text_location(&Direction::Left);
+            self.delete();
+        }
     }
 
     fn delete(&mut self) {
@@ -81,7 +90,7 @@ impl View {
         let grapheme_delta = new_len.saturating_sub(old_len);
         if grapheme_delta > 0 {
             // si la nouvelle longueur est plus grande
-            self.move_right();
+            self.move_text_location(&Direction::Right);
         }
         self.needs_redraw = true;
     }
